fix: 修复 SpeakerDiarizerTest 音频 WAV 解码测试并移除 .broken 隔离源
- pcmToWavBytes 编码辅助运算符优先级修正：(int) Math.max(...) * 32767 会先截断为 0，导致全部采样点编码为 0，testDecode16bitWav 第 3 采样点偏差超限
- writeShort 写入 audioFormat 时 1 改为 (short)1（int 不能赋值给 short）
- 将此前被隔离（quarantine）为 .broken 的两个测试源恢复为 .java

+++ utils-support-deeplearning-parent/utils-support-deeplearning-onnx-starter/src/test/java/com/chua/deeplearning/support/onnx/audio/SpeakerDiarizerTest.java
@@ -347,7 +347,7 @@ class SpeakerDiarizerTest {
         writeStr(wav, 8, "WAVE");
         writeStr(wav, 12, "fmt ");
         writeInt(wav, 16, 16);
-        writeShort(wav, 20, 1);
+        writeShort(wav, 20, (short) 1);
         writeShort(wav, 22, (short) numChannels);
         writeInt(wav, 24, sampleRate);
         writeInt(wav, 28, byteRate);
@@ -357,7 +357,7 @@ class SpeakerDiarizerTest {
         writeInt(wav, 40, dataSize);
         int off = 44;
         for (float s : samples) {
-            int val = (int) Math.max(-1.0f, Math.min(1.0f, s)) * 32767;
+            int val = (int) (Math.max(-1.0f, Math.min(1.0f, s)) * 32767);
             wav[off++] = (byte) (val & 0xff);
             wav[off++] = (byte) ((val >> 8) & 0xff);
         }
